+++ pyproj/_compat.pyx
@@ -1,0 +1,7 @@
+from pyproj.compat import pystrdecode
+
+
+cdef cstrdecode(const char *instring):
+    if instring != NULL:
+        return pystrdecode(instring)
+    return None